Improve make checks, fix annoyances
Without this patch there are a number of annoyances with the make
checks.  This patch fixes the following:

 1. Checks execute against .git/ and .terraform/ directories
 2. terraform validate executes redundantly

The patch refactors make.sh to use a general find | xargs pattern with a
single function to manage exclusion lists.

check_shebangs is removed because `#! /bin/bash -e` should not be used
because it does not take effect if the script is called via bash
<filename>.  Instead, calling `set -e` is preferred.