Fixup - Associate a service account with the gsutil instance
Without this patch the instance used for positive behavior tests of the gsutil
command does not have read access to the bucket fixture.  This causes failures.

This patch fixes the problem by associating the CI service account with the
instance.  A future improvement could introduce a new SA with more limited
access, e.g. get only.  Negative tests to cover retry behavior may also be
introduced in a future improvement.